Mr.Xia 会员服务-会员管理 完善添加会员 2019-04-25 18:50

+++ commons/src/main/java/org/fresh/gd/commons/consts/pojo/dto/vip/GdAddVipLvDTO.java
@@ -27,7 +27,7 @@ public class GdAddVipLvDTO {
      */
     @NotEmpty
     @ApiModelProperty("会员折扣")
-    private Integer vipdiscount;
+    private Double vipdiscount;
 
     /**
      * 所需积分
